experiments/requirements_experiments.txt
hydra-core
hydra-colorlog
rich
matplotlib
seaborn
pandas
wandb
coax
dm-haiku
jax[cuda,cpu]
hydra-submitit-launcher